Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe

Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe

Utilisation

cartofriches.friches(
  coddep = NULL,
  code_insee = NULL,
  lon_lat = NULL,
  in_bbox = NULL,
  ordering = NULL,
  fields = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL,
  ...
)

Arguments

coddep

Codes INSEE des départements.

code_insee

Codes INSEE communaux ou des arrondissements municipaux.

lon_lat

Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)

in_bbox

Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

surface_max

Surface maximale de l'unité foncière

surface_min

Surface minimale de l'unité foncière

urba_zone_type

Type de zone d'urbanisme

...

pour futurs developpements

Valeur de retour

Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un dataframe

Exemples

cartofriches.friches(code_insee="59350")
#> # A tibble: 155 × 14
#>    site_id     site_nom          site_type site_statut comm_nom comm_insee dep  
#>    <chr>       <chr>             <chr>     <chr>       <chr>    <chr>      <chr>
#>  1 59350_10190 Teinturerie       inconnu   friche pot… LILLE    59350      59   
#>  2 59350_10192 WILVIA MEUBLEX    inconnu   friche pot… LILLE    59350      59   
#>  3 59350_10209 Atelier de mécan… inconnu   friche pot… LILLE    59350      59   
#>  4 59350_10210 Station Service … inconnu   friche pot… LILLE    59350      59   
#>  5 59350_10211 Assitance et Ins… inconnu   friche pot… LILLE    59350      59   
#>  6 59350_10795 Construction de … inconnu   inconnu     LILLE    59350      59   
#>  7 59350_1102  Atelier de const… inconnu   friche pot… LILLE    59350      59   
#>  8 59350_11724 Appel à projet R… inconnu   friche ave… LILLE    59350      59   
#>  9 59350_11777 Cie Européenne d… inconnu   inconnu     LILLE    59350      59   
#> 10 59350_11787 Appel à projet R… inconnu   friche ave… LILLE    59350      59   
#> # ℹ 145 more rows
#> # ℹ 7 more variables: proprio_personne <chr>, unite_fonciere_surface <dbl>,
#> #   unite_fonciere_refcad <chr>, source_nom <chr>, proprio_type_lib <list>,
#> #   source_nature <chr>, urba_zone_type <chr>
cartofriches.friches(coddep="59")
#> ■■■■■■■■■■■                       35% | ETA:  0s
#> ■■■■■■■■■■■■■■■■■■■■■■            70% | ETA:  0s
#> # A tibble: 1,580 × 14
#>    site_id     site_nom          site_type site_statut comm_nom comm_insee dep  
#>    <chr>       <chr>             <chr>     <chr>       <chr>    <chr>      <chr>
#>  1 59002_10038 carrière des Plo… inconnu   friche pot… ABSCON   59002      59   
#>  2 59002_3873  Atelier de répar… inconnu   friche pot… ABSCON   59002      59   
#>  3 59005_4472  Distillerie       inconnu   friche pot… ALLENNE… 59005      59   
#>  4 59007_13298 l'Air Liquide     inconnu   inconnu     ANHIERS  59007      59   
#>  5 59008_10042 Expanver          inconnu   friche pot… ANICHE   59008      59   
#>  6 59008_10218 Le Mont-Rouge     inconnu   friche pot… ANICHE   59008      59   
#>  7 59008_10220 SARL MIROITERIE … inconnu   friche pot… ANICHE   59008      59   
#>  8 59008_10234 Terril Saint-Lou… inconnu   friche pot… ANICHE   59008      59   
#>  9 59008_10236 Usine d'agglomér… inconnu   friche pot… ANICHE   59008      59   
#> 10 59008_10560 59008_10560       inconnu   inconnu     ANICHE   59008      59   
#> # ℹ 1,570 more rows
#> # ℹ 7 more variables: proprio_personne <chr>, unite_fonciere_surface <dbl>,
#> #   unite_fonciere_refcad <chr>, source_nom <chr>, proprio_type_lib <list>,
#> #   source_nature <chr>, urba_zone_type <chr>
cartofriches.friches(in_bbox=c(3, 50, 4, 51))
#> ■■■■■■■■■■■■■                     40% | ETA:  0s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         79% | ETA:  0s
#> # A tibble: 1,393 × 14
#>    site_id     site_nom          site_type site_statut comm_nom comm_insee dep  
#>    <chr>       <chr>             <chr>     <chr>       <chr>    <chr>      <chr>
#>  1 02103_21221 "Appel à projet … inconnu   friche ave… BOUE     02103      02   
#>  2 02352_15846 "Carosserie Indu… inconnu   friche pot… GOUY     02352      02   
#>  3 02760_15858 "Décharge DIB"    inconnu   friche pot… LA VALL… 02760      02   
#>  4 59002_10038 "carrière des Pl… inconnu   friche pot… ABSCON   59002      59   
#>  5 59002_3873  "Atelier de répa… inconnu   friche pot… ABSCON   59002      59   
#>  6 59007_13298 "l'Air Liquide"   inconnu   inconnu     ANHIERS  59007      59   
#>  7 59008_10042 "Expanver"        inconnu   friche pot… ANICHE   59008      59   
#>  8 59008_10218 "Le Mont-Rouge"   inconnu   friche pot… ANICHE   59008      59   
#>  9 59008_10220 "SARL MIROITERIE… inconnu   friche pot… ANICHE   59008      59   
#> 10 59008_10234 "Terril Saint-Lo… inconnu   friche pot… ANICHE   59008      59   
#> # ℹ 1,383 more rows
#> # ℹ 7 more variables: proprio_personne <chr>, unite_fonciere_surface <dbl>,
#> #   unite_fonciere_refcad <chr>, source_nom <chr>, proprio_type_lib <list>,
#> #   source_nature <chr>, urba_zone_type <chr>