Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe

Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe

Utilisation

cartofriches.geofriches(
  coddep = NULL,
  code_insee = NULL,
  lon_lat = NULL,
  in_bbox = NULL,
  ordering = NULL,
  fields = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL,
  ...
)

Arguments

coddep

Codes INSEE des départements.

code_insee

Codes INSEE communaux ou des arrondissements municipaux.

lon_lat

Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)

in_bbox

Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

surface_max

Surface maximale de l'unité foncière

surface_min

Surface minimale de l'unité foncière

urba_zone_type

Type de zone d'urbanisme

...

pour futurs developpements

Valeur de retour

Retourne les friches issues de Cartofriches pour le périmètre demandé sous forme d’un geodataframe

Exemples

cartofriches.geofriches(code_insee="59350")
#> Simple feature collection with 135 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.9743 ymin: 50.6082 xmax: 3.1134 ymax: 50.6547
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>             id
#> 1  59350_10190
#> 2  59350_10192
#> 3  59350_10209
#> 4  59350_10210
#> 5  59350_10211
#> 6  59350_10795
#> 7   59350_1102
#> 8  59350_11724
#> 9  59350_11777
#> 10 59350_11787
#>                                                                                                                                                              site_nom
#> 1                                                                                                                                                         Teinturerie
#> 2                                                                                                                                                      WILVIA MEUBLEX
#> 3                                                                                                                                                Atelier de mécanique
#> 4                                                                                                                                                Station Service Esso
#> 5                                                                                                                                  Assitance et Inspection Techniques
#> 6                                                                                                                                              Construction de Wagons
#> 7                                                                                                                                 Atelier de constructions mécaniques
#> 8  Appel à projet Recyclage Foncier 2021 2nde édition (Dossier N° 6057240) - Hauts-de-France - 59 - Nord - Lille (59260) - friche: LOMME EURATECHNOLOGIE - ILOT 28 29
#> 9                                                                                   Cie Européenne d'accumulateurs (SA), anc. TUDOR (Cie Générale d'Electricité) CEAC
#> 10           Appel à projet Recyclage Foncier 2021 2nde édition (Dossier N° 5781164) - Hauts-de-France - 59 - Nord - Lille (59000) - friche: 93-95 rue Balzac - Lille
#>    site_type site_adresse        site_statut comm_nom comm_insee dep
#> 1    inconnu          { } friche potentielle    LILLE      59350  59
#> 2    inconnu          { } friche potentielle    LILLE      59350  59
#> 3    inconnu          { } friche potentielle    LILLE      59350  59
#> 4    inconnu          { } friche potentielle    LILLE      59350  59
#> 5    inconnu          { } friche potentielle    LILLE      59350  59
#> 6    inconnu          { }            inconnu    LILLE      59350  59
#> 7    inconnu          { } friche potentielle    LILLE      59350  59
#> 8    inconnu          { } friche avec projet    LILLE      59350  59
#> 9    inconnu          { }            inconnu    LILLE      59350  59
#> 10   inconnu          { } friche avec projet    LILLE      59350  59
#>     proprio_personne unite_fonciere_surface
#> 1    personne morale                 234.17
#> 2    personne morale                 234.17
#> 3    personne morale                 234.17
#> 4    personne morale                 234.17
#> 5    personne morale                 234.17
#> 6    personne morale                 122.92
#> 7  personne physique                 248.12
#> 8    personne morale                5390.18
#> 9    personne morale               67896.90
#> 10   personne morale                1891.81
#>                                                                                                                                                                                                               unite_fonciere_refcad
#> 1                                                                                                                                                                                                                    59350000IK0025
#> 2                                                                                                                                                                                                                    59350000IK0025
#> 3                                                                                                                                                                                                                    59350000IK0025
#> 4                                                                                                                                                                                                                    59350000IK0025
#> 5                                                                                                                                                                                                                    59350000IK0025
#> 6                                                                                                                                                                                                                    59350000DV0166
#> 7                                                                                                                                                                       59350000OX0165|59350000OX0166|59350000OX0167|59350000OX0176
#> 8                                                                                                                                                                                                     593503550C6992|593503550C6993
#> 9                                                                                            59350000DK0124|59350000DK0125|59350000DK0126|59350000DK0326|59350000DK0329|59350000DK0332|59350000DK0341|59350000DK0342|59350000DK0508
#> 10 59350000DR0324|59350000DR0325|59350000DR0326|59350000DR0329|59350000DR0330|59350000DR0331|59350000DR0332|59350000DR0333|59350000DR0334|59350000DR0335|59350000DR0336|59350000DR0337|59350000DR0338|59350000DR0339|59350000DR0340
#>                                                 source_nom   source_nature
#> 1           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 2           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 3           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 4           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 5           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 6       Site Basias ou Basol vérifié par le Cerema en 2020             MTE
#> 7           Site Basias ou Basol non vérifié par le Cerema MTE non vérifié
#> 8                             Appel à projet Fonds Friches     fond friche
#> 9  Friche étude Ademe 2021, intéressante pour du PV au sol           Ademe
#> 10                            Appel à projet Fonds Friches     fond friche
#>    urba_zone_type                 proprio_type_lib
#> 1               U                      DEPARTEMENT
#> 2               U                      DEPARTEMENT
#> 3               U                      DEPARTEMENT
#> 4               U                      DEPARTEMENT
#> 5               U                      DEPARTEMENT
#> 6               U SCI - SOCIETE CIVILE IMMOBILIERE
#> 7               U                PERSONNE PHYSIQUE
#> 8               U          SEM OU SPLA INDETERMINE
#> 9               U         ETABLISSEMENT INDUSTRIEL
#> 10              U                 INTERCOMMUNALITE
#>                          geometry
#> 1  MULTIPOLYGON (((3.0744 50.6...
#> 2  MULTIPOLYGON (((3.0744 50.6...
#> 3  MULTIPOLYGON (((3.0744 50.6...
#> 4  MULTIPOLYGON (((3.0744 50.6...
#> 5  MULTIPOLYGON (((3.0744 50.6...
#> 6  MULTIPOLYGON (((3.0401 50.6...
#> 7  MULTIPOLYGON (((3.0671 50.6...
#> 8  MULTIPOLYGON (((3.019 50.63...
#> 9  MULTIPOLYGON (((3.0553 50.6...
#> 10 MULTIPOLYGON (((3.0465 50.6...
cartofriches.geofriches(coddep="59")
#> ■■■■■■■■■■■                       35% | ETA: 18s
#> ■■■■■■■■■■■■■■■■■■■■■■            70% | ETA:  6s
#> Simple feature collection with 1429 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.1216 ymin: 49.9853 xmax: 4.1477 ymax: 51.0623
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>             id                     site_nom site_type site_adresse
#> 1  59002_10038          carrière des Plombs   inconnu          { }
#> 2   59002_3873        Atelier de réparation   inconnu          { }
#> 3   59005_4472                  Distillerie   inconnu          { }
#> 4  59007_13298                l'Air Liquide   inconnu          { }
#> 5  59008_10042                     Expanver   inconnu          { }
#> 6  59008_10218                Le Mont-Rouge   inconnu          { }
#> 7  59008_10220 SARL MIROITERIE SAINT MARTIN   inconnu          { }
#> 8  59008_10234       Terril Saint-Louis 126   inconnu          { }
#> 9  59008_10236           Usine d'agglomérés   inconnu          { }
#> 10 59008_10560                  59008_10560   inconnu          { }
#>           site_statut            comm_nom comm_insee dep  proprio_personne
#> 1  friche potentielle              ABSCON      59002  59   personne morale
#> 2  friche potentielle              ABSCON      59002  59 personne physique
#> 3  friche potentielle ALLENNES-LES-MARAIS      59005  59 personne physique
#> 4             inconnu             ANHIERS      59007  59   personne morale
#> 5  friche potentielle              ANICHE      59008  59   personne morale
#> 6  friche potentielle              ANICHE      59008  59   personne morale
#> 7  friche potentielle              ANICHE      59008  59   personne morale
#> 8  friche potentielle              ANICHE      59008  59   personne morale
#> 9  friche potentielle              ANICHE      59008  59   personne morale
#> 10            inconnu              ANICHE      59008  59   personne morale
#>    unite_fonciere_surface
#> 1               103894.44
#> 2                 4287.23
#> 3                  704.62
#> 4                74328.57
#> 5                39623.75
#> 6               151864.71
#> 7                 2074.03
#> 8               537518.11
#> 9               537518.11
#> 10              537518.11
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       unite_fonciere_refcad
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               59002000AN0033|59002000AN0173|59002000AN0174|59002000AN0181
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              59002000AL0066|59002000AL0394|59002000AL0430
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            590050000B3022
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            59178000AE0158
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    59008000AA0027|59008000AA0028|59008000AC0005|59008000AC0006|59008000AC0022|59008000AC0259|59008000AC0423|59008000AC0425|59008000AC0642
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  591920000U0093|591920000U0094|591920000U0095|591920000U0096|591920000U0097|591920000U0098|591920000U0534
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            59008000AH0874
#> 8  595740000B0950|595740000B0951|595740000B0952|595740000B0953|595740000B0954|595740000B0955|595740000B0956|595740000B0958|595740000B0959|595740000B0960|595740000B0961|595740000B0963|595740000B0964|595740000B0965|595740000B0966|595740000B0967|595740000B0969|595740000B0970|595740000B0971|595740000B0972|595740000B0973|595740000B0974|595740000B1013|595740000B1016|595740000B1017|595740000B1019|595740000B1020|595740000B1021|595740000B1022|595740000B1028|595740000B1029|595740000B1031|595740000B1032|595740000B1033|595740000B1034|595740000B1035|595740000B1036|595740000B1037|595740000B1038|595740000B1039|595740000B1040|595740000B1041|595740000B1042|595740000B1043|595740000B1044|595740000B1045|595740000B1790|595740000B1816|595740000B1817|595740000B1818|595740000B1819|595740000B1820|595740000B1821|595740000B2359|595740000B2727|595740000B2891|595740000B3026|595740000B3119|595740000B3123|595740000B3186|595740000B3198|595740000B3200|595740000B3204|595740000B3208|595740000B3209|595740000B3230|595740000B3232|595740000B3233|595740000B3240|595740000B3256|595740000B3261|595740000B3274|595740000B3275|595740000B3276|595740000B3277|595740000B3278|595740000B3279|595740000B3280|595740000B3281|595740000B3283|595740000B3284|595740000B3287|595740000B3288|595740000B3289|595740000B3330|595740000B3331|595740000B3351
#> 9  595740000B0950|595740000B0951|595740000B0952|595740000B0953|595740000B0954|595740000B0955|595740000B0956|595740000B0958|595740000B0959|595740000B0960|595740000B0961|595740000B0963|595740000B0964|595740000B0965|595740000B0966|595740000B0967|595740000B0969|595740000B0970|595740000B0971|595740000B0972|595740000B0973|595740000B0974|595740000B1013|595740000B1016|595740000B1017|595740000B1019|595740000B1020|595740000B1021|595740000B1022|595740000B1028|595740000B1029|595740000B1031|595740000B1032|595740000B1033|595740000B1034|595740000B1035|595740000B1036|595740000B1037|595740000B1038|595740000B1039|595740000B1040|595740000B1041|595740000B1042|595740000B1043|595740000B1044|595740000B1045|595740000B1790|595740000B1816|595740000B1817|595740000B1818|595740000B1819|595740000B1820|595740000B1821|595740000B2359|595740000B2727|595740000B2891|595740000B3026|595740000B3119|595740000B3123|595740000B3186|595740000B3198|595740000B3200|595740000B3204|595740000B3208|595740000B3209|595740000B3230|595740000B3232|595740000B3233|595740000B3240|595740000B3256|595740000B3261|595740000B3274|595740000B3275|595740000B3276|595740000B3277|595740000B3278|595740000B3279|595740000B3280|595740000B3281|595740000B3283|595740000B3284|595740000B3287|595740000B3288|595740000B3289|595740000B3330|595740000B3331|595740000B3351
#> 10 595740000B0950|595740000B0951|595740000B0952|595740000B0953|595740000B0954|595740000B0955|595740000B0956|595740000B0958|595740000B0959|595740000B0960|595740000B0961|595740000B0963|595740000B0964|595740000B0965|595740000B0966|595740000B0967|595740000B0969|595740000B0970|595740000B0971|595740000B0972|595740000B0973|595740000B0974|595740000B1013|595740000B1016|595740000B1017|595740000B1019|595740000B1020|595740000B1021|595740000B1022|595740000B1028|595740000B1029|595740000B1031|595740000B1032|595740000B1033|595740000B1034|595740000B1035|595740000B1036|595740000B1037|595740000B1038|595740000B1039|595740000B1040|595740000B1041|595740000B1042|595740000B1043|595740000B1044|595740000B1045|595740000B1790|595740000B1816|595740000B1817|595740000B1818|595740000B1819|595740000B1820|595740000B1821|595740000B2359|595740000B2727|595740000B2891|595740000B3026|595740000B3119|595740000B3123|595740000B3186|595740000B3198|595740000B3200|595740000B3204|595740000B3208|595740000B3209|595740000B3230|595740000B3232|595740000B3233|595740000B3240|595740000B3256|595740000B3261|595740000B3274|595740000B3275|595740000B3276|595740000B3277|595740000B3278|595740000B3279|595740000B3280|595740000B3281|595740000B3283|595740000B3284|595740000B3287|595740000B3288|595740000B3289|595740000B3330|595740000B3331|595740000B3351
#>                                        source_nom
#> 1  Site Basias ou Basol non vérifié par le Cerema
#> 2  Site Basias ou Basol non vérifié par le Cerema
#> 3  Site Basias ou Basol non vérifié par le Cerema
#> 4                         Friche étude Ademe 2021
#> 5  Site Basias ou Basol non vérifié par le Cerema
#> 6  Site Basias ou Basol non vérifié par le Cerema
#> 7  Site Basias ou Basol non vérifié par le Cerema
#> 8  Site Basias ou Basol non vérifié par le Cerema
#> 9  Site Basias ou Basol non vérifié par le Cerema
#> 10                        Friche étude Ademe 2021
#>                                              proprio_type_lib   source_nature
#> 1                                           [ "DEPARTEMENT" ] MTE non vérifié
#> 2                                     [ "PERSONNE PHYSIQUE" ] MTE non vérifié
#> 3                                     [ "PERSONNE PHYSIQUE" ] MTE non vérifié
#> 4                              [ "ETABLISSEMENT INDUSTRIEL" ]           MTE\n
#> 5  [ "SNCF – SOCIETE NATIONALE DES CHEMINS DE FER FRANCAIS" ] MTE non vérifié
#> 6                              [ "ETABLISSEMENT INDUSTRIEL" ] MTE non vérifié
#> 7                                               [ "COMMUNE" ] MTE non vérifié
#> 8                                      [ "INTERCOMMUNALITE" ] MTE non vérifié
#> 9                                      [ "INTERCOMMUNALITE" ] MTE non vérifié
#> 10                                     [ "INTERCOMMUNALITE" ]           MTE\n
#>    urba_zone_type                       geometry
#> 1             { } MULTIPOLYGON (((3.3082 50.3...
#> 2             { } MULTIPOLYGON (((3.3 50.3312...
#> 3             { } MULTIPOLYGON (((2.9545 50.5...
#> 4               N MULTIPOLYGON (((3.1491 50.4...
#> 5               A MULTIPOLYGON (((3.2608 50.3...
#> 6               U MULTIPOLYGON (((3.2583 50.3...
#> 7               U MULTIPOLYGON (((3.2446 50.3...
#> 8               U MULTIPOLYGON (((3.2648 50.3...
#> 9               U MULTIPOLYGON (((3.2648 50.3...
#> 10              A MULTIPOLYGON (((3.2675 50.3...
cartofriches.geofriches(in_bbox=c(3, 50, 4, 51))
#> ■■■■■■■■■■■■■                     40% | ETA: 11s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         79% | ETA:  3s
#> Simple feature collection with 1261 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.9718 ymin: 50.0021 xmax: 4.0032 ymax: 50.7944
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>             id
#> 1  02103_21221
#> 2  02352_15846
#> 3  02760_15858
#> 4  59002_10038
#> 5   59002_3873
#> 6  59007_13298
#> 7  59008_10042
#> 8  59008_10218
#> 9  59008_10220
#> 10 59008_10234
#>                                                                                             site_nom
#> 1  Appel à projet Recyclage Foncier 2021 (Dossier N° 3592780) - Hauts-de-France - 02 - Aisne - Boué 
#> 2                                                       Carosserie Industrielle de l'Escaut ; TABARY
#> 3                                                                                       Décharge DIB
#> 4                                                                                carrière des Plombs
#> 5                                                                              Atelier de réparation
#> 6                                                                                      l'Air Liquide
#> 7                                                                                           Expanver
#> 8                                                                                      Le Mont-Rouge
#> 9                                                                       SARL MIROITERIE SAINT MARTIN
#> 10                                                                            Terril Saint-Louis 126
#>    site_type site_adresse        site_statut          comm_nom comm_insee dep
#> 1    inconnu          { } friche avec projet              BOUE      02103  02
#> 2    inconnu          { } friche potentielle              GOUY      02352  02
#> 3    inconnu          { } friche potentielle LA VALLEE-MULATRE      02760  02
#> 4    inconnu          { } friche potentielle            ABSCON      59002  59
#> 5    inconnu          { } friche potentielle            ABSCON      59002  59
#> 6    inconnu          { }            inconnu           ANHIERS      59007  59
#> 7    inconnu          { } friche potentielle            ANICHE      59008  59
#> 8    inconnu          { } friche potentielle            ANICHE      59008  59
#> 9    inconnu          { } friche potentielle            ANICHE      59008  59
#> 10   inconnu          { } friche potentielle            ANICHE      59008  59
#>     proprio_personne unite_fonciere_surface
#> 1    personne morale               17432.97
#> 2    personne morale                6040.32
#> 3  personne physique               33547.43
#> 4    personne morale              103894.44
#> 5  personne physique                4287.23
#> 6    personne morale               74328.57
#> 7    personne morale               39623.75
#> 8    personne morale              151864.71
#> 9    personne morale                2074.03
#> 10   personne morale              537518.11
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       unite_fonciere_refcad
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       02103000AC0472|02103000AC0473|02103000AC0474|02103000AC0475|02103000AC0476|02103000AC0483|02103000AC0484|02103000AC0485|02103000AC0486|02103000AC0487|02103000AC0488|02103000AC0489
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 023520000A0475|023520000A0481|023520000A0482|023520000A0483|023520000A0540|02352000ZL0011
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             02760000ZD0032|02760000ZD0033
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               59002000AN0033|59002000AN0173|59002000AN0174|59002000AN0181
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              59002000AL0066|59002000AL0394|59002000AL0430
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            59178000AE0158
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    59008000AA0027|59008000AA0028|59008000AC0005|59008000AC0006|59008000AC0022|59008000AC0259|59008000AC0423|59008000AC0425|59008000AC0642
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  591920000U0093|591920000U0094|591920000U0095|591920000U0096|591920000U0097|591920000U0098|591920000U0534
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            59008000AH0874
#> 10 595740000B0950|595740000B0951|595740000B0952|595740000B0953|595740000B0954|595740000B0955|595740000B0956|595740000B0958|595740000B0959|595740000B0960|595740000B0961|595740000B0963|595740000B0964|595740000B0965|595740000B0966|595740000B0967|595740000B0969|595740000B0970|595740000B0971|595740000B0972|595740000B0973|595740000B0974|595740000B1013|595740000B1016|595740000B1017|595740000B1019|595740000B1020|595740000B1021|595740000B1022|595740000B1028|595740000B1029|595740000B1031|595740000B1032|595740000B1033|595740000B1034|595740000B1035|595740000B1036|595740000B1037|595740000B1038|595740000B1039|595740000B1040|595740000B1041|595740000B1042|595740000B1043|595740000B1044|595740000B1045|595740000B1790|595740000B1816|595740000B1817|595740000B1818|595740000B1819|595740000B1820|595740000B1821|595740000B2359|595740000B2727|595740000B2891|595740000B3026|595740000B3119|595740000B3123|595740000B3186|595740000B3198|595740000B3200|595740000B3204|595740000B3208|595740000B3209|595740000B3230|595740000B3232|595740000B3233|595740000B3240|595740000B3256|595740000B3261|595740000B3274|595740000B3275|595740000B3276|595740000B3277|595740000B3278|595740000B3279|595740000B3280|595740000B3281|595740000B3283|595740000B3284|595740000B3287|595740000B3288|595740000B3289|595740000B3330|595740000B3331|595740000B3351
#>                                        source_nom
#> 1                    Appel à projet Fonds Friches
#> 2  Site Basias ou Basol non vérifié par le Cerema
#> 3  Site Basias ou Basol non vérifié par le Cerema
#> 4  Site Basias ou Basol non vérifié par le Cerema
#> 5  Site Basias ou Basol non vérifié par le Cerema
#> 6                         Friche étude Ademe 2021
#> 7  Site Basias ou Basol non vérifié par le Cerema
#> 8  Site Basias ou Basol non vérifié par le Cerema
#> 9  Site Basias ou Basol non vérifié par le Cerema
#> 10 Site Basias ou Basol non vérifié par le Cerema
#>                                              proprio_type_lib   source_nature
#> 1                                               [ "COMMUNE" ]     fond friche
#> 2                      [ "SCI - SOCIETE CIVILE IMMOBILIERE" ] MTE non vérifié
#> 3                                     [ "PERSONNE PHYSIQUE" ] MTE non vérifié
#> 4                                           [ "DEPARTEMENT" ] MTE non vérifié
#> 5                                     [ "PERSONNE PHYSIQUE" ] MTE non vérifié
#> 6                              [ "ETABLISSEMENT INDUSTRIEL" ]           MTE\n
#> 7  [ "SNCF – SOCIETE NATIONALE DES CHEMINS DE FER FRANCAIS" ] MTE non vérifié
#> 8                              [ "ETABLISSEMENT INDUSTRIEL" ] MTE non vérifié
#> 9                                               [ "COMMUNE" ] MTE non vérifié
#> 10                                     [ "INTERCOMMUNALITE" ] MTE non vérifié
#>    urba_zone_type                       geometry
#> 1             AUc MULTIPOLYGON (((3.6991 50.0...
#> 2             { } MULTIPOLYGON (((3.2578 50.0...
#> 3               A MULTIPOLYGON (((3.5521 50.0...
#> 4             { } MULTIPOLYGON (((3.3082 50.3...
#> 5             { } MULTIPOLYGON (((3.3 50.3312...
#> 6               N MULTIPOLYGON (((3.1491 50.4...
#> 7               A MULTIPOLYGON (((3.2608 50.3...
#> 8               U MULTIPOLYGON (((3.2583 50.3...
#> 9               U MULTIPOLYGON (((3.2446 50.3...
#> 10              U MULTIPOLYGON (((3.2648 50.3...