Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf

Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf

Utilisation

dvf.geomutations(
  code_insee = NULL,
  lon_lat = NULL,
  in_bbox = NULL,
  ordering = NULL,
  fields = NULL,
  anneemut = NULL,
  anneemut_max = NULL,
  anneemut_min = NULL,
  codtypbien = NULL,
  codtypproa = NULL,
  codtypprov = NULL,
  filtre = NULL,
  idnatmut = NULL,
  sbati_max = NULL,
  sbati_min = NULL,
  segmtab = NULL,
  sterr_max = NULL,
  sterr_min = NULL,
  valeurfonc_max = NULL,
  valeurfonc_min = NULL,
  vefa = NULL,
  ...
)

Arguments

code_insee

Codes INSEE communaux ou des arrondissements municipaux.

lon_lat

Coordonnée du point au sein de la ou des friches renvoyées c(longitude, latitude)

in_bbox

Emprise rectangulaire sous la forme d’un vecteur c(longitude_min, latitude_min, longitude_min, latitude_max)

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

anneemut

Annee de la mutation (>= 2010) - cf anneemut

anneemut_max

Annee de la mutation maximale - cf anneemut

anneemut_min

Annee de la mutation minimale - cf anneemut

codtypbien

Code(s) de la typologie de bien à sélectionner (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypbien=11,121 - cf codtypbien

codtypproa

Code(s) de la typologie de l'acheteur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypproa=P,F7 - cf codtypproa

codtypprov

Code(s) de la typologie du vendeur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypprov=P,F7 - cf codtypprov

filtre

Code alphanumerique permettant d'exclure des transactions particulières - cf filtre

idnatmut

Code(s) de nature de mutation (il est possible d'en demander plusieurs en séparant par une virgule - exemple : idnatmut=1,2 - cf idnatmut

sbati_max

Surface maximale bâtie vendue - cf sbati

sbati_min

Surface minimale bâtie vendue - cf sbati

segmtab

Note de segment du terrain à bâtir - cf segmtab

sterr_max

Surface maximale de terrain vendue - cf sterr

sterr_min

Surface minimale de terrain vendue - cf sterr

valeurfonc_max

Valeur foncière maximale (€) - cf valeurfonc

valeurfonc_min

Valeur foncière minimale (€) - cf valeurfonc

vefa

Vente en l'état futur d'achèvement - cf vefa

...

pour futurs developpements

Valeur de retour

Retourne les mutations issues de DVF+ opendata pour le périmètre demandée sous forme d’un sf

Exemples

dvf.geomutations(code_insee="59001")
#> Simple feature collection with 84 features and 21 fields (with 1 geometry empty)
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 3.1742 ymin: 50.1593 xmax: 3.305 ymax: 50.257
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>                          idmutinvar                       idopendata    datemut
#> 1  274db1bcf6991ca07cfe0f641ca5c7b8 274db1bcf6991ca07cfe0f641ca5c7b8 2022-08-05
#> 2  b05d9e0e9848cf966341470c061f1553 b05d9e0e9848cf966341470c061f1553 2022-03-16
#> 3  950ab57f428fabd06957b87fc7469cbd 950ab57f428fabd06957b87fc7469cbd 2022-09-16
#> 4  7f3aaf2d7274e2b330211ea5392d83fa 7f3aaf2d7274e2b330211ea5392d83fa 2022-10-25
#> 5  7fca9a8e78fb13f2e95c21746da2fb40 7fca9a8e78fb13f2e95c21746da2fb40 2022-10-26
#> 6  6faba458b70702c27ed945f808a7b232 6faba458b70702c27ed945f808a7b232 2022-03-04
#> 7  5b6706487711534638482b57063b0693 5b6706487711534638482b57063b0693 2021-05-17
#> 8  06874211e786c2bf66ea8d344bee491a 06874211e786c2bf66ea8d344bee491a 2021-12-28
#> 9  5e5f60095de46a2b2d2c8bfb3abdedf8 5e5f60095de46a2b2d2c8bfb3abdedf8 2021-03-03
#> 10 2d561a1e58b22bf8048986170e61e857 2d561a1e58b22bf8048986170e61e857 2021-02-26
#>    anneemut coddep libnatmut  vefa valeurfonc nbcomm nbpar nbparmut   sterr
#> 1      2022     59     Vente FALSE   47000.00      1     3        3  766.00
#> 2      2022     59     Vente FALSE  280000.00      1     1        1 1790.00
#> 3      2022     59     Vente FALSE  182100.00      1     3        3  601.00
#> 4      2022     59     Vente FALSE   70000.00      1     4        4  925.00
#> 5      2022     59     Vente FALSE  280000.00      1     2        2 2675.00
#> 6      2022     59     Vente FALSE  276000.00      1     1        1  271.00
#> 7      2021     59     Vente FALSE  309600.00      1     1        1  710.00
#> 8      2021     59     Vente FALSE    8500.00      1     1        1 6070.00
#> 9      2021     59     Vente FALSE  289000.00      1     1        1  711.00
#> 10     2021     59     Vente FALSE  225000.00      1     1        1  870.00
#>    nbvolmut nblocmut
#> 1         0        0
#> 2         0        1
#> 3         0        1
#> 4         0        5
#> 5         0        3
#> 6         0        2
#> 7         0        1
#> 8         0        0
#> 9         0        1
#> 10        0        1
#>                                                                                                     l_idlocmut
#> 1                                                                                                          { }
#> 2                                                                                      [ "590010010516868_0" ]
#> 3                                                                                      [ "590010010530297_0" ]
#> 4  [ "590010010532997_0", "590010010532998_0", "590010010532999_0", "590010010533000_0", "590010010533001_0" ]
#> 5                                            [ "590010010534250_0", "590010010534251_0", "590010010534252_0" ]
#> 6                                                                 [ "590010010516777_0", "590010010516778_0" ]
#> 7                                                                                      [ "590010010653272_0" ]
#> 8                                                                                                          { }
#> 9                                                                                      [ "590010010652529_0" ]
#> 10                                                                                     [ "590010010652269_0" ]
#>     sbati codtypbien                   libtypbien l_codinsee
#> 1    0.00         21          TERRAIN DE TYPE TAB      59001
#> 2  102.00        111                   UNE MAISON      59001
#> 3  111.00        111                   UNE MAISON      59001
#> 4  144.00        111                   UNE MAISON      59001
#> 5  149.00        111                   UNE MAISON      59001
#> 6  122.00        111                   UNE MAISON      59001
#> 7  174.00        111                   UNE MAISON      59001
#> 8    0.00       2313 TERRAIN DE TYPE TERRE ET PRE      59001
#> 9  154.00        111                   UNE MAISON      59001
#> 10 110.00        111                   UNE MAISON      59001
#>                                                           l_idpar
#> 1                  590010000U1585, 590010000U1587, 590010000U1588
#> 2                                                  59001000ZL0084
#> 3                  590010000U1384, 590010000U1519, 590010000U1575
#> 4  590010000U1221, 590010000U1222, 590010000U1514, 590010000U1515
#> 5                                  590010000U1331, 590010000U1572
#> 6                                                  590010000U1524
#> 7                                                  590010000U1345
#> 8                                                  59001000ZK0019
#> 9                                                  59001000ZL0140
#> 10                                                 59001000ZL0145
#>                                                        l_idparmut
#> 1                  590010000U1585, 590010000U1587, 590010000U1588
#> 2                                                  59001000ZL0084
#> 3                  590010000U1384, 590010000U1519, 590010000U1575
#> 4  590010000U1221, 590010000U1222, 590010000U1514, 590010000U1515
#> 5                                  590010000U1331, 590010000U1572
#> 6                                                  590010000U1524
#> 7                                                  590010000U1345
#> 8                                                  59001000ZK0019
#> 9                                                  59001000ZL0140
#> 10                                                 59001000ZL0145
#>                          geometry
#> 1  MULTIPOLYGON (((3.2148 50.2...
#> 2  MULTIPOLYGON (((3.2128 50.2...
#> 3  MULTIPOLYGON (((3.2109 50.2...
#> 4  MULTIPOLYGON (((3.2153 50.2...
#> 5  MULTIPOLYGON (((3.2108 50.2...
#> 6  MULTIPOLYGON (((3.2123 50.2...
#> 7  MULTIPOLYGON (((3.2122 50.2...
#> 8  MULTIPOLYGON (((3.225 50.24...
#> 9  MULTIPOLYGON (((3.2087 50.2...
#> 10 MULTIPOLYGON (((3.2097 50.2...
dvf.geomutations(in_bbox=c(3, 50, 3.01, 50.01), fields="all")
#> Simple feature collection with 12 features and 60 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.6145 ymin: 49.7138 xmax: 3.129 ymax: 50.0439
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>                          idmutinvar                       idopendata idnatmut
#> 1  7a6624df246ed63710d2b1a3b1e78be1 7a6624df246ed63710d2b1a3b1e78be1        1
#> 2  939c71ca867f8ff97b7224afa850daab 939c71ca867f8ff97b7224afa850daab        1
#> 3  e2c3cf92693785f6c2875f7467e44fca e2c3cf92693785f6c2875f7467e44fca        1
#> 4  0ac94d2733cdff35e12c1a052250e248 0ac94d2733cdff35e12c1a052250e248        1
#> 5  cc731bbc2352304739ec37464dc66b05 cc731bbc2352304739ec37464dc66b05        1
#> 6  482fe5ad9cd9995c7fc5da77e6dd6cf2 482fe5ad9cd9995c7fc5da77e6dd6cf2        1
#> 7  269e96af3969e662204a7ccf24f81457 269e96af3969e662204a7ccf24f81457        1
#> 8  1b17471fdbe03bd4b0e41fb3876cefbe 1b17471fdbe03bd4b0e41fb3876cefbe        1
#> 9  aa1dc25fca84a8543d2687166708b1ba aa1dc25fca84a8543d2687166708b1ba        1
#> 10 052564a8611a514a521ed4dc29f22584 052564a8611a514a521ed4dc29f22584        1
#>    codservch refdoc    datemut anneemut moismut coddep libnatmut nbartcgi
#> 1        { }    { } 2022-07-23     2022       7     80     Vente        0
#> 2        { }    { } 2021-12-13     2021      12     80     Vente        0
#> 3        { }    { } 2021-06-16     2021       6     80     Vente        0
#> 4        { }    { } 2020-12-16     2020      12     80     Vente        0
#> 5        { }    { } 2019-05-20     2019       5     80     Vente        0
#> 6        { }    { } 2018-05-16     2018       5     80     Vente        0
#> 7        { }    { } 2018-10-11     2018      10     80     Vente        0
#> 8        { }    { } 2015-04-23     2015       4     80     Vente        0
#> 9        { }    { } 2015-04-17     2015       4     80     Vente        0
#> 10       { }    { } 2015-07-28     2015       7     80     Vente        0
#>    l_artcgi  vefa valeurfonc nbdispo nblot nbcomm nbsection nbpar nbparmut
#> 1       { } FALSE   33785.00       1     0      3         3     5        5
#> 2       { } FALSE   56887.00       1     0      3         3    12       12
#> 3       { } FALSE  114000.00       2     2      2         2     3        2
#> 4       { } FALSE  208919.00       1     0      2         4     5        5
#> 5       { } FALSE   93000.00       2     0      2         2     2        2
#> 6       { } FALSE  210862.00       1     0      3         6    13       13
#> 7       { } FALSE   64040.00       2     0      2         2     3        3
#> 8       { } FALSE   83500.00       2     0      2         2     2        2
#> 9       { } FALSE   18525.00       1     0      1         1     1        1
#> 10      { } FALSE  111000.00       2     0      2         2     4        4
#>    nbsuf     sterr nbvolmut nblocmut
#> 1      5  52096.00        0        0
#> 2     14  72752.00        0        0
#> 3      2   2153.00        0        1
#> 4      5  87853.00        0        0
#> 5      2    449.00        0        2
#> 6     17 392057.00        0        0
#> 7      3   9958.00        0        0
#> 8      3    709.00        0        1
#> 9      1  18250.00        0        0
#> 10     5   4849.00        0        1
#>                                      l_idlocmut nblocmai nblocapt nblocdep
#> 1                                           { }        0        0        0
#> 2                                           { }        0        0        0
#> 3                       [ "800160015462080_0" ]        1        0        0
#> 4                                           { }        0        0        0
#> 5  [ "801290015702728_0", "806770015702729_0" ]        2        0        0
#> 6                                           { }        0        0        0
#> 7                                           { }        0        0        0
#> 8                       [ "806200016016880_0" ]        1        0        0
#> 9                                           { }        0        0        0
#> 10                      [ "800470016017184_0" ]        1        0        0
#>    nblocact nbapt1pp nbapt2pp nbapt3pp nbapt4pp nbapt5pp nbmai1pp nbmai2pp
#> 1         0        0        0        0        0        0        0        0
#> 2         0        0        0        0        0        0        0        0
#> 3         0        0        0        0        0        0        0        0
#> 4         0        0        0        0        0        0        0        0
#> 5         0        0        0        0        0        0        0        0
#> 6         0        0        0        0        0        0        0        0
#> 7         0        0        0        0        0        0        0        0
#> 8         0        0        0        0        0        0        0        0
#> 9         0        0        0        0        0        0        0        0
#> 10        0        0        0        0        0        0        0        0
#>    nbmai3pp nbmai4pp nbmai5pp  sbati sbatmai sbatapt sbatact sapt1pp sapt2pp
#> 1         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 2         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 3         1        0        0  59.00   59.00    0.00    0.00    0.00    0.00
#> 4         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 5         1        1        0 173.00  173.00    0.00    0.00    0.00    0.00
#> 6         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 7         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 8         0        0        1  92.00   92.00    0.00    0.00    0.00    0.00
#> 9         0        0        0   0.00    0.00    0.00    0.00    0.00    0.00
#> 10        1        0        0  60.00   60.00    0.00    0.00    0.00    0.00
#>    sapt3pp sapt4pp sapt5pp smai1pp smai2pp smai3pp smai4pp smai5pp codtypbien
#> 1     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 2     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 3     0.00    0.00    0.00    0.00    0.00   59.00    0.00    0.00        111
#> 4     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 5     0.00    0.00    0.00    0.00    0.00  120.00   53.00    0.00        112
#> 6     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 7     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 8     0.00    0.00    0.00    0.00    0.00    0.00    0.00   92.00        111
#> 9     0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00       2313
#> 10    0.00    0.00    0.00    0.00    0.00   60.00    0.00    0.00        111
#>                      libtypbien          l_codinsee              l_section
#> 1  TERRAIN DE TYPE TERRE ET PRE 80014, 80275, 80552             0A, 0T, ZD
#> 2  TERRAIN DE TYPE TERRE ET PRE 80014, 80475, 80601             0B, 0X, 0Z
#> 3                    UNE MAISON        80016, 80284                 0F, AP
#> 4  TERRAIN DE TYPE TERRE ET PRE        80298, 80601         0X, 0Z, ZB, ZN
#> 5                   DES MAISONS        80129, 80677                 AB, AC
#> 6  TERRAIN DE TYPE TERRE ET PRE 80438, 80552, 80747 0A, 0B, 0C, 0S, ZW, ZX
#> 7  TERRAIN DE TYPE TERRE ET PRE        80342, 80601                 0X, ZR
#> 8                    UNE MAISON        80601, 80620                 0A, BE
#> 9  TERRAIN DE TYPE TERRE ET PRE               80601                     0Z
#> 10                   UNE MAISON        80047, 80435                 0A, AI
#>                                                                                                                                                                                                           l_idpar
#> 1                                                                                                                                  800140000A0003, 80275000ZD0002, 805520000T0021, 805520000T0052, 805520000T0053
#> 2                  800140000B0006, 804750000X0106, 804750000X0107, 806010000X0016, 806010000X0026, 806010000X0027, 806010000X0102, 806010000X0103, 806010000X0135, 806010000X0191, 806010000Z0032, 806010000Z0093
#> 3                                                                                                                                                                  80016000AP0164, 802840000F0434, 802840000F0435
#> 4                                                                                                                                  80298000ZN0005, 806010000X0097, 806010000Z0036, 806010000Z0154, 80601000ZB0001
#> 5                                                                                                                                                                                  80129000AB0080, 80677000AC0406
#> 6  80438000ZW0027, 80438000ZX0012, 805520000S0136, 807470000A0053, 807470000B0011, 807470000B0012, 807470000B0046, 807470000B0069, 807470000B0071, 807470000B0094, 807470000B0103, 807470000C0071, 807470000C0072
#> 7                                                                                                                                                                  80342000ZR0027, 80342000ZR0028, 806010000X0087
#> 8                                                                                                                                                                                  806010000A0544, 80620000BE0111
#> 9                                                                                                                                                                                                  806010000Z0151
#> 10                                                                                                                                                 80047000AI0031, 804350000A0399, 804350000A0404, 804350000A0406
#>                                                                                                                                                                                                        l_idparmut
#> 1                                                                                                                                  800140000A0003, 80275000ZD0002, 805520000T0021, 805520000T0052, 805520000T0053
#> 2                  800140000B0006, 804750000X0106, 804750000X0107, 806010000X0016, 806010000X0026, 806010000X0027, 806010000X0102, 806010000X0103, 806010000X0135, 806010000X0191, 806010000Z0032, 806010000Z0093
#> 3                                                                                                                                                                                  802840000F0434, 802840000F0435
#> 4                                                                                                                                  80298000ZN0005, 806010000X0097, 806010000Z0036, 806010000Z0154, 80601000ZB0001
#> 5                                                                                                                                                                                  80129000AB0080, 80677000AC0406
#> 6  80438000ZW0027, 80438000ZX0012, 805520000S0136, 807470000A0053, 807470000B0011, 807470000B0012, 807470000B0046, 807470000B0069, 807470000B0071, 807470000B0094, 807470000B0103, 807470000C0071, 807470000C0072
#> 7                                                                                                                                                                  80342000ZR0027, 80342000ZR0028, 806010000X0087
#> 8                                                                                                                                                                                  806010000A0544, 80620000BE0111
#> 9                                                                                                                                                                                                  806010000Z0151
#> 10                                                                                                                                                 80047000AI0031, 804350000A0399, 804350000A0404, 804350000A0406
#>                                                                                  l_dcnt
#> 1      52096.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 2      72752.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 3       0.00, 2153.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 4  69294.00, 18559.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 5        0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 449.00
#> 6  389522.00, 0.00, 0.00, 0.00, 0.00, 2535.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 7       9958.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 8      0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 387.00, 0.00, 0.00, 0.00, 322.00
#> 9      18250.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00
#> 10   2818.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 2031.00
#>                          geometry
#> 1  MULTIPOLYGON (((3.0243 49.9...
#> 2  MULTIPOLYGON (((3.0631 49.9...
#> 3  MULTIPOLYGON (((2.6384 49.9...
#> 4  MULTIPOLYGON (((3.0088 50.0...
#> 5  MULTIPOLYGON (((3.1025 49.9...
#> 6  MULTIPOLYGON (((3.0197 49.9...
#> 7  MULTIPOLYGON (((3.0336 50.0...
#> 8  MULTIPOLYGON (((2.9359 49.9...
#> 9  MULTIPOLYGON (((3.0035 50.0...
#> 10 MULTIPOLYGON (((2.6545 50.0...
dvf.geomutations(code_insee=c("59350", "59646"), valeurfonc_min=1000000)
#> ■■■■■■■■■■■■■■■■■                 52% | ETA: 12s
#> Simple feature collection with 1075 features and 21 fields (with 9 geometries empty)
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 2.9115 ymin: 50.5116 xmax: 3.2153 ymax: 50.7728
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>                          idmutinvar                       idopendata    datemut
#> 1  dc9eb9044ab0e6f340f7e9668b497adb dc9eb9044ab0e6f340f7e9668b497adb 2022-09-29
#> 2  478ebde8f40d13393c255af2f87c6919 478ebde8f40d13393c255af2f87c6919 2022-06-15
#> 3  7064cf16d65a1e0eba5547a6b552b743 7064cf16d65a1e0eba5547a6b552b743 2022-04-04
#> 4  7071cdcd24c76886c83e5785868f0eba 7071cdcd24c76886c83e5785868f0eba 2022-06-24
#> 5  ec2ec3a1351929cd60f60d27f96a1503 ec2ec3a1351929cd60f60d27f96a1503 2022-10-07
#> 6  cb6fa844dfa5bd8e3d78d5ab509faa01 cb6fa844dfa5bd8e3d78d5ab509faa01 2022-04-28
#> 7  97333f246947b39cb5cbdf5cc138e44e 97333f246947b39cb5cbdf5cc138e44e 2022-10-07
#> 8  29d824cc34de32d1c94dc79717bfb2b7 29d824cc34de32d1c94dc79717bfb2b7 2022-05-19
#> 9  9db8536ac4a2fb3af771e053cee34698 9db8536ac4a2fb3af771e053cee34698 2022-04-07
#> 10 58742fb22c28133f52c58ef0a4e0878e 58742fb22c28133f52c58ef0a4e0878e 2022-09-30
#>    anneemut coddep libnatmut  vefa  valeurfonc nbcomm nbpar nbparmut
#> 1      2022     59     Vente FALSE 10800000.00      1     2        1
#> 2      2022     59     Vente FALSE  1180000.00      1     2        1
#> 3      2022     59     Vente FALSE  1200000.00      1     1        0
#> 4      2022     59     Vente FALSE  3139440.00      1     5        5
#> 5      2022     59     Vente FALSE  2280000.00      1     1        0
#> 6      2022     59     Vente FALSE  1320000.00      1     1        0
#> 7      2022     59     Vente FALSE  3830000.00      1     2        0
#> 8      2022     59     Vente FALSE  2850000.00      1     1        1
#> 9      2022     59     Vente FALSE  2300000.00      1     1        1
#> 10     2022     59     Vente FALSE 17600000.00      1     1        1
#>                                                                                      l_idparmut
#> 1                                                                          [ "59350000TM0035" ]
#> 2                                                                          [ "59350000KR0107" ]
#> 3                                                                                           { }
#> 4  [ "59350000CK0172", "59350000CK0173", "59350000CK0174", "59350000CK0191", "59350000CK0207" ]
#> 5                                                                                           { }
#> 6                                                                                           { }
#> 7                                                                                           { }
#> 8                                                                          [ "59350000LS0220" ]
#> 9                                                                          [ "59350000LN0056" ]
#> 10                                                                         [ "59350000LM0177" ]
#>      sterr nbvolmut nblocmut
#> 1  1095.00        0        5
#> 2   280.00        0        2
#> 3     0.00        0        5
#> 4   471.00        0        0
#> 5     0.00        0        2
#> 6     0.00        0        1
#> 7     0.00        0       13
#> 8   407.00        0        1
#> 9   171.00        0        6
#> 10 1236.00        0       11
#>                                                                                                                                                                                                                                                                             l_idlocmut
#> 1                                                                                                                                                                          [ "593500010490348_0", "593500010490349_0", "593500010490350_0", "593500010490351_0", "593500010490352_0" ]
#> 2                                                                                                                                                                                                                                         [ "593500010473590_0", "593500010473591_0" ]
#> 3                                                                                                                                                                          [ "593500010465825_0", "593500010465826_0", "593500010465827_0", "593500010465828_0", "593500010465829_0" ]
#> 4                                                                                                                                                                                                                                                                                  { }
#> 5                                                                                                                                                                                                                                         [ "593500010492537_0", "593500010492538_0" ]
#> 6                                                                                                                                                                                                                                                              [ "593500010466412_0" ]
#> 7  [ "593500010490479_0", "593500010490480_0", "593500010490481_0", "593500010490482_0", "593500010490483_0", "593500010490484_0", "593500010490485_0", "593500010490486_0", "593500010490487_0", "593500010490488_0", "593500010490489_0", "593500010490490_0", "593500010490491_0" ]
#> 8                                                                                                                                                                                                                                                              [ "593500010470056_0" ]
#> 9                                                                                                                                                     [ "593500010466326_0", "593500010466327_0", "593500010466328_0", "593500010466329_0", "593500010466330_0", "593500010466331_0" ]
#> 10                                           [ "593500010489186_0", "593500010489187_0", "593500010489188_0", "593500010489189_0", "593500010489190_0", "593500010489191_0", "593500010489192_0", "593500010489193_0", "593500010489194_0", "593500010489195_0", "593500010489196_0" ]
#>      sbati codtypbien                     libtypbien l_codinsee
#> 1  3844.00         14                       ACTIVITE      59350
#> 2   343.00        111                     UNE MAISON      59350
#> 3   308.00        152 BATI MIXTE - LOGEMENT/ACTIVITE      59350
#> 4     0.00        229    TERRAIN ARTIFICIALISE MIXTE      59350
#> 5   544.00         14                       ACTIVITE      59350
#> 6   234.00         14                       ACTIVITE      59350
#> 7   910.00         14                       ACTIVITE      59350
#> 8   930.00         14                       ACTIVITE      59350
#> 9   546.00        152 BATI MIXTE - LOGEMENT/ACTIVITE      59350
#> 10 4575.00         14                       ACTIVITE      59350
#>                                                                           l_idpar
#> 1                                                  59350000TM0026, 59350000TM0035
#> 2                                                  59350000KR0107, 59350000NW0044
#> 3                                                                  59350000SW0001
#> 4  59350000CK0172, 59350000CK0173, 59350000CK0174, 59350000CK0191, 59350000CK0207
#> 5                                                                  59350000NX0143
#> 6                                                                  59350000LN0014
#> 7                                                  59350000LR0314, 59350000NX0143
#> 8                                                                  59350000LS0220
#> 9                                                                  59350000LN0056
#> 10                                                                 59350000LM0177
#>                          geometry
#> 1  MULTIPOLYGON (((3.0692 50.6...
#> 2  MULTIPOLYGON (((3.0515 50.6...
#> 3  MULTIPOLYGON (((3.0474 50.6...
#> 4  MULTIPOLYGON (((3.0924 50.6...
#> 5  MULTIPOLYGON (((3.0593 50.6...
#> 6  MULTIPOLYGON (((3.0653 50.6...
#> 7  MULTIPOLYGON (((3.0623 50.6...
#> 8  MULTIPOLYGON (((3.0636 50.6...
#> 9  MULTIPOLYGON (((3.0657 50.6...
#> 10 MULTIPOLYGON (((3.0666 50.6...