Retourne les indicateurs annuels ou triennaux de prix à l’aire d’attraction de la ville

Retourne les indicateurs annuels ou triennaux de prix à l’aire d’attraction de la ville

Utilisation

prix.aav(code_insee, periode = "annuel", annee = NULL, ordering = NULL, ...)

Arguments

code_insee

Code AAV de l'aire INSEE

periode

Prend la valeur « annuel » ou « triennal ». Defaults to annuel.

annee

Année de mutation

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

...

pour futurs developpements

Valeur de retour

dataframe: données sur les marchés immobiliers

Exemples

prix.aav(code_insee="001")
#> # A tibble: 13 × 617
#>    annee aav   libaav nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>         <dbl>               <dbl>        <dbl>
#>  1 2010  001   Paris        200934        61221967990.         7729
#>  2 2011  001   Paris        181418        60879056997.         7951
#>  3 2012  001   Paris        159961        57145321506.         8194
#>  4 2013  001   Paris        158539        54375723547.         7795
#>  5 2014  001   Paris        152913        59392592699.         7269
#>  6 2015  001   Paris        173360        61864083537.         6958
#>  7 2016  001   Paris        184861        67224027302.         7228
#>  8 2017  001   Paris        215107        81227156256.         8022
#>  9 2018  001   Paris        211338        83212479870.         8084
#> 10 2019  001   Paris        229964        95848474329.         9022
#> 11 2020  001   Paris        206415        90433412917.         9159
#> 12 2021  001   Paris        231230       101677745332.        13620
#> 13 2022  001   Paris        224424       105202106579.        13295
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.aav(code_insee=c("001", "002"), annee="2020")
#> # A tibble: 2 × 613
#>   annee aav   libaav nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>   <chr> <chr> <chr>         <dbl>               <dbl>        <dbl>
#> 1 2020  001   Paris        206415        90433412917.         9159
#> 2 2020  002   Lyon          38654        12529107485.         3460
#> # ℹ 607 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.aav(code_insee=c("001", "002"), periode="triennal")
#> # A tibble: 22 × 617
#>    annee aav   libaav nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>         <dbl>               <dbl>        <dbl>
#>  1 2011  001   Paris        542313       179246346493.        23874
#>  2 2012  001   Paris        499918       172400102050.        23940
#>  3 2013  001   Paris        471413       170913637752.        23258
#>  4 2014  001   Paris        484812       175632399783.        22022
#>  5 2015  001   Paris        511134       188480703538.        21455
#>  6 2016  001   Paris        573328       210315267096.        22208
#>  7 2017  001   Paris        611306       231663663428.        23334
#>  8 2018  001   Paris        656409       260288110455.        25128
#>  9 2019  001   Paris        647717       269494367116.        26265
#> 10 2020  001   Paris        667609       287959632578.        31801
#> # ℹ 12 more rows
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>, …