Retourne les indicateurs annuels ou triennaux de prix à la commune

Retourne les indicateurs annuels ou triennaux de prix à la commune

Utilisation

prix.communes(
  code_insee,
  periode = "annuel",
  annee = NULL,
  ordering = NULL,
  ...
)

Arguments

code_insee

Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)

periode

Prend la valeur « annuel » ou « triennal ». Defaults to annuel.

annee

Année de mutation

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

...

pour futurs developpements

Valeur de retour

dataframe: données sur les marchés immobiliers

Exemples

prix.communes(code_insee="59350")
#> # A tibble: 13 × 573
#>    annee codgeo libgeo nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr>  <chr>         <dbl>               <dbl>        <dbl>
#>  1 2010  59350  Lille          5120          837659942.           57
#>  2 2011  59350  Lille          4474          882766180.           52
#>  3 2012  59350  Lille          4200          813426374.           78
#>  4 2013  59350  Lille          3971          781613569.           45
#>  5 2014  59350  Lille          3931          804964493.           40
#>  6 2015  59350  Lille          4472          913893573.           36
#>  7 2016  59350  Lille          4708         1016164793.           43
#>  8 2017  59350  Lille          5387         1315158597.           49
#>  9 2018  59350  Lille          5587         1533348525.           50
#> 10 2019  59350  Lille          5914         1316236006.           45
#> 11 2020  59350  Lille          5115         1338281971.           52
#> 12 2021  59350  Lille          5638         1368565071.           74
#> 13 2022  59350  Lille          5188         1793537471.           44
#> # ℹ 567 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1113 <dbl>,
#> #   valeurfonc_sum_cod1113 <dbl>, valeurfonc_q25_cod1113 <dbl>, …
prix.communes(code_insee=c("59350", "59646"), annee="2020")
#> # A tibble: 2 × 519
#>   annee codgeo libgeo    nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>   <chr> <chr>  <chr>            <dbl>               <dbl>        <dbl>
#> 1 2020  59350  Lille             5115         1338281971.           52
#> 2 2020  59646  Wasquehal          365          129589305.           11
#> # ℹ 513 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1113 <dbl>,
#> #   valeurfonc_sum_cod1113 <dbl>, valeurfonc_q25_cod1113 <dbl>, …
prix.communes(code_insee=c("59350", "59646"), periode="triennal")
#> # A tibble: 22 × 603
#>    annee codgeo libgeo nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr>  <chr>         <dbl>               <dbl>        <dbl>
#>  1 2011  59350  Lille         13794         2533852496.          187
#>  2 2012  59350  Lille         12645         2477806123.          175
#>  3 2013  59350  Lille         12102         2400004436.          163
#>  4 2014  59350  Lille         12374         2500471635.          121
#>  5 2015  59350  Lille         13111         2735022859.          119
#>  6 2016  59350  Lille         14567         3245216964.          128
#>  7 2017  59350  Lille         15682         3864671916.          142
#>  8 2018  59350  Lille         16888         4164743129.          144
#>  9 2019  59350  Lille         16616         4187866503.          147
#> 10 2020  59350  Lille         16667         4023083048.          171
#> # ℹ 12 more rows
#> # ℹ 597 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>, …