Retourne les indicateurs annuels ou triennaux de prix au département

Retourne les indicateurs annuels ou triennaux de prix au département

Utilisation

prix.departements(
  coddep,
  periode = "annuel",
  annee = NULL,
  ordering = NULL,
  ...
)

Arguments

coddep

Codes des départements

periode

Prend la valeur « annuel » ou « triennal ». Defaults to annuel.

annee

Année de mutation

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

...

pour futurs developpements

Valeur de retour

dataframe: données sur les marchés immobiliers

Exemples

prix.departements(coddep="59")
#> # A tibble: 13 × 617
#>    annee dep   libdep nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>         <dbl>               <dbl>        <dbl>
#>  1 2010  59    Nord          34222         5714013800.         3715
#>  2 2011  59    Nord          31832         5559759012.         3814
#>  3 2012  59    Nord          29013         5126548658.         3865
#>  4 2013  59    Nord          28614         4935566738.         3742
#>  5 2014  59    Nord          26677         4565658940.         3343
#>  6 2015  59    Nord          30673         5717041435.         3020
#>  7 2016  59    Nord          32922         5985091937.         3375
#>  8 2017  59    Nord          36359         7182136188.         3273
#>  9 2018  59    Nord          36736         7224403888.         3492
#> 10 2019  59    Nord          40044         7574858097.         3747
#> 11 2020  59    Nord          38368         7621794304.         4034
#> 12 2021  59    Nord          43956         9913070183.         5607
#> 13 2022  59    Nord          42336        11152937517.         6061
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.departements(coddep=c("59", "62"), annee="2020")
#> # A tibble: 2 × 613
#>   annee dep   libdep        nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>   <chr> <chr> <chr>                <dbl>               <dbl>        <dbl>
#> 1 2020  59    Nord                 38368         7621794304.         4034
#> 2 2020  62    Pas-de-Calais        19386         3230603292.         3769
#> # ℹ 607 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.departements(coddep=c("59", "62"), periode="triennal")
#> # A tibble: 22 × 617
#>    annee dep   libdep nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>         <dbl>               <dbl>        <dbl>
#>  1 2011  59    Nord          95067        16400321471.        11394
#>  2 2012  59    Nord          89459        15621874408.        11421
#>  3 2013  59    Nord          84304        14627774336.        10950
#>  4 2014  59    Nord          85964        15218267113.        10105
#>  5 2015  59    Nord          90272        16267792312.         9738
#>  6 2016  59    Nord          99954        18884269560.         9668
#>  7 2017  59    Nord         106017        20391632013.        10140
#>  8 2018  59    Nord         113139        21981398173.        10512
#>  9 2019  59    Nord         115148        22421056289.        11273
#> 10 2020  59    Nord         122368        25109722584.        13388
#> # ℹ 12 more rows
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>, …