Retourne les indicateurs annuels ou triennaux de prix à l'EPCI

Retourne les indicateurs annuels ou triennaux de prix à l'EPCI

Utilisation

prix.epci(code_insee, periode = "annuel", annee = NULL, ordering = NULL, ...)

Arguments

code_insee

Codes des EPCI

periode

Prend la valeur « annuel » ou « triennal ». Defaults to annuel.

annee

Année de mutation

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

...

pour futurs developpements

Valeur de retour

dataframe: données sur les marchés immobiliers

Exemples

prix.epci(code_insee="200093201")
#> # A tibble: 13 × 617
#>    annee epci      libepci         nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr>     <chr>                  <dbl>               <dbl>        <dbl>
#>  1 2010  200093201 Métropole Euro…        19204         3539032732.          886
#>  2 2011  200093201 Métropole Euro…        17763         3500800814.          868
#>  3 2012  200093201 Métropole Euro…        16119         3278008917.          948
#>  4 2013  200093201 Métropole Euro…        15669         3090838287.          815
#>  5 2014  200093201 Métropole Euro…        14574         2884974564.          655
#>  6 2015  200093201 Métropole Euro…        17073         3881853205.          593
#>  7 2016  200093201 Métropole Euro…        18410         3907619208.          683
#>  8 2017  200093201 Métropole Euro…        20250         4894000621.          652
#>  9 2018  200093201 Métropole Euro…        20488         4855194688.          734
#> 10 2019  200093201 Métropole Euro…        22017         4913083378.          808
#> 11 2020  200093201 Métropole Euro…        20579         4938937195.          812
#> 12 2021  200093201 Métropole Euro…        23042         6203621574.         1203
#> 13 2022  200093201 Métropole Euro…        22736         7647413855.         1252
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.epci(code_insee="200093201", periode="triennal")
#> # A tibble: 11 × 617
#>    annee epci      libepci         nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr>     <chr>                  <dbl>               <dbl>        <dbl>
#>  1 2011  200093201 Métropole Euro…        53086        10317842462.         2702
#>  2 2012  200093201 Métropole Euro…        49551         9869648018          2631
#>  3 2013  200093201 Métropole Euro…        46362         9253821768.         2418
#>  4 2014  200093201 Métropole Euro…        47316         9857666056.         2063
#>  5 2015  200093201 Métropole Euro…        50057        10674446977.         1931
#>  6 2016  200093201 Métropole Euro…        55733        12683473034.         1928
#>  7 2017  200093201 Métropole Euro…        59148        13656814517.         2069
#>  8 2018  200093201 Métropole Euro…        62755        14662278687.         2194
#>  9 2019  200093201 Métropole Euro…        63084        14707215261.         2354
#> 10 2020  200093201 Métropole Euro…        65638        16055642148.         2823
#> 11 2021  200093201 Métropole Euro…        66357        18789972624.         3267
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …