Retourne les indicateurs annuels ou triennaux de prix à la région

Retourne les indicateurs annuels ou triennaux de prix à la région

Utilisation

prix.regions(
  code_insee,
  periode = "annuel",
  annee = NULL,
  ordering = NULL,
  ...
)

Arguments

code_insee

Codes des régions

periode

Prend la valeur « annuel » ou « triennal ». Defaults to annuel.

annee

Année de mutation

ordering

Champs à utiliser pour ordonner le résultat. Default NULL

...

pour futurs developpements

Valeur de retour

dataframe: données sur les marchés immobiliers

Exemples

prix.regions(code_insee="32")
#> # A tibble: 13 × 617
#>    annee reg   libreg          nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>                  <dbl>               <dbl>        <dbl>
#>  1 2010  32    Hauts-de-France        70402        11292758526.        12614
#>  2 2011  32    Hauts-de-France        65816        10944357163.        12957
#>  3 2012  32    Hauts-de-France        59645         9785681692.        13042
#>  4 2013  32    Hauts-de-France        59882         9629088229.        12493
#>  5 2014  32    Hauts-de-France        56638        14072133331.        11533
#>  6 2015  32    Hauts-de-France        64708        10919657525.        11053
#>  7 2016  32    Hauts-de-France        69499        11501345732.        11419
#>  8 2017  32    Hauts-de-France        77710        13604840622.        11704
#>  9 2018  32    Hauts-de-France        77931        13822000993.        12350
#> 10 2019  32    Hauts-de-France        85524        15097871646.        13227
#> 11 2020  32    Hauts-de-France        85265        15403789227.        13496
#> 12 2021  32    Hauts-de-France        97633        19724632437.        19707
#> 13 2022  32    Hauts-de-France        87029        19725296664.        18051
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.regions(code_insee=c("32", "11"), annee="2020")
#> # A tibble: 2 × 613
#>   annee reg   libreg          nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>   <chr> <chr> <chr>                  <dbl>               <dbl>        <dbl>
#> 1 2020  32    Hauts-de-France        85265        15403789227.        13496
#> 2 2020  11    Île-de-France         193818        87892734421.         6899
#> # ℹ 607 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>,
#> #   valeurfonc_sum_cod1111 <dbl>, valeurfonc_q25_cod1111 <dbl>, …
prix.regions(code_insee=c("32", "11"), periode="triennal")
#> # A tibble: 22 × 617
#>    annee reg   libreg          nbtrans_cod1 valeurfonc_sum_cod1 nbtrans_cod2
#>    <chr> <chr> <chr>                  <dbl>               <dbl>        <dbl>
#>  1 2011  32    Hauts-de-France       195863        32022797382.        38613
#>  2 2012  32    Hauts-de-France       185343        30359127085.        38492
#>  3 2013  32    Hauts-de-France       176165        33486903252.        37068
#>  4 2014  32    Hauts-de-France       181228        34620879085.        35079
#>  5 2015  32    Hauts-de-France       190845        36493136588.        34005
#>  6 2016  32    Hauts-de-France       211917        36025843879.        34176
#>  7 2017  32    Hauts-de-France       225140        38928187347.        35473
#>  8 2018  32    Hauts-de-France       241165        42524713261.        37281
#>  9 2019  32    Hauts-de-France       248720        44323661866.        39073
#> 10 2020  32    Hauts-de-France       268422        50226293310.        46430
#> # ℹ 12 more rows
#> # ℹ 611 more variables: valeurfonc_sum_cod2 <dbl>, nbtrans_cod11 <dbl>,
#> #   valeurfonc_sum_cod11 <dbl>, nbtrans_cod111 <dbl>,
#> #   valeurfonc_sum_cod111 <dbl>, valeurfonc_q25_cod111 <dbl>,
#> #   valeurfonc_median_cod111 <dbl>, valeurfonc_q75_cod111 <dbl>,
#> #   pxm2_q25_cod111 <dbl>, pxm2_median_cod111 <dbl>, pxm2_q75_cod111 <dbl>,
#> #   sbati_sum_cod111 <dbl>, sbati_median_cod111 <dbl>, nbtrans_cod1111 <dbl>, …